src/requirements.txt
bidict==0.21.2
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
dnspython==1.16.0
eventlet==0.31.1
firebase-admin==5.0.1
Flask==2.0.1
Flask-SocketIO==5.1.0
Flask-SSLify==0.1.5
flask-talisman==0.8.1
gcloud==0.18.3
gevent==21.1.2
google-api-core==1.31.0
google-api-python-client==2.14.1
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwcrypto==0.9.1
MarkupSafe==2.0.1
msgpack==1.0.2
oauth2client==4.1.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-engineio==4.2.0
python-jwt==3.3.0
python-socketio==5.3.0
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
